requirements.txt
# Flask==2.0.1
# mysql-connector-python==8.0.27
# Werkzeug==2.2.2

flask
mysql-connector-python
